Clean up CLRF delimited stream
Changed to take advantage of `ends_with` and used `mem::swap` to return
the vec that was built during iteration rather than copying and then
clearing. I've also cleaned up the pattern matching. The debug lines
were removed, as they can be handled at the call site if they are
needed. The only case that loses information is `Ok(Ready(None))`, but
as discussed elsewhere that case is actually signalling EOF so it should
be refactored out